+++ src/0x.ts
@@ -172,6 +172,7 @@ export class ZeroEx {
         (this.tokenRegistry as any)._invalidateContractInstance();
         await (this.token as any)._invalidateContractInstancesAsync();
         (this.proxy as any)._invalidateContractInstance();
+        (this.etherToken as any)._invalidateContractInstance();
     }
     /**
      * Get user Ethereum addresses available through the supplied web3 instance available for sending transactions.