+++ CHANGELOG.md
@@ -6,6 +6,11 @@ v0.8.0 - TBD
     * Update contract artifacts to reflect latest changes to the smart contracts (0xproject/contracts#59)
     * Add `zeroEx.proxy.isAuthorizedAsync` and `zeroEx.proxy.getAuthorizedAddressesAsync` (#89)
     * Add `zeroEx.token.subscribeAsync` (#90)
+    * Make contract invalidation functions private (#90)
+        * `zeroEx.token.invalidateContractInstancesAsync`
+        * `zeroEx.exchange.invalidateContractInstancesAsync`
+        * `zeroEx.proxy.invalidateContractInstance`
+        * `zeroEx.tokenRegistry.invalidateContractInstance`
 
 v0.7.1 - _Jun. 26, 2017_
 ------------------------