+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # CHANGELOG
 
-v0.8.0 - TBD
+v0.8.0 - _Jul. 4, 2017_
 ------------------------
     * Add the ability to call methods on different authorized versions of the Exchange smart contract (#82)
     * Update contract artifacts to reflect latest changes to the smart contracts (0xproject/contracts#59)